magiceightball/requirements.txt
asgiref==3.5.0
autopep8==1.6.0
dj-database-url==0.5.0
Django==4.0.3
django-cors-headers==3.11.0
django-filter==1.1.0
django-heroku==0.3.1
django-mongodb-engine==0.6.0
djangorestframework==3.13.1
djangorestframework-filters==0.11.1
djangotoolbox==1.8.0
djongo==1.3.6
dnspython==2.2.1
fuzzywuzzy==0.18.0
gunicorn==20.1.0
mongoengine==0.24.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pycountry==22.3.5
pymongo==3.12.3
python-dotenv==0.20.0
python-Levenshtein==0.12.2
pytz==2022.1
sqlparse==0.2.4
toml==0.10.2
whitenoise==6.0.0
